项目开发对java语言的支持完成，剩下功能在git上更新

+++ src/main/java/com/yupi/jieOj/model/dto/question/JudgeConfig.java
@@ -8,14 +8,14 @@ public class JudgeConfig {
     /**
      * 时间限(ms)
      */
-    private long timeLimit;
+    private int timeLimit;
     /**
      * 内存限制(KB)
      */
-    private long memoryLimit;
+    private int memoryLimit;
 
     /**
      * 堆栈限制(KB)
      */
-    private long stackLimit;
+    private int stackLimit;
 }
